Refatorando a classe model.
close #80
