+++ .flake8
@@ -1,3 +1,4 @@
 [flake8]
 max-line-length = 120
 exclude=.venv, devpro/encurtador/migrations
+per-file-ignores = __init__.py:F401
