+++ devpro/encurtador/models/__init__.py
@@ -1,0 +1,2 @@
+from devpro.encurtador.models.admin import User, UserManager
+from devpro.encurtador.models.url import UrlRedirect, UrlLog
