+++ devpro/encurtador/models/admin.py
@@ -4,7 +4,6 @@ from django.core.mail import send_mail
 from django.utils.translation import gettext_lazy as _
 from django.db import models
 from django.utils import timezone
-from django.urls import reverse
 
 
 class UserManager(BaseUserManager):
@@ -93,37 +92,3 @@ class User(AbstractBaseUser, PermissionsMixin):
     def email_user(self, subject, message, from_email=None, **kwargs):
         """Send an email to this user."""
         send_mail(subject, message, from_email, [self.email], **kwargs)
-
-
-class UrlRedirect(models.Model):
-    """
-    Classe que armazena as urls.
-
-    Campos
-    destino : guarda a url real
-    slug: o apelido da url
-    criado_em : momento que a url foi criada
-    atualizado_em : momento que a url foi modifica
-    """
-    destino = models.URLField(max_length=1024)
-    slug = models.SlugField(max_length=128, unique=True)
-    criado_em = models.DateTimeField(auto_now_add=True)
-    atualizado_em = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return f'{self.slug} -> {self.destino}'
-
-    def get_absolute_url(self):
-        return reverse('relatorio_name', kwargs={'slug': self.slug})
-
-
-class UrlLog(models.Model):
-    criado_em = models.DateTimeField(auto_now_add=True)
-    origem = models.URLField(max_length=1024, null=True, blank=True)
-    user_agent = models.CharField(max_length=512, null=True, blank=True)
-    host = models.CharField(max_length=512, null=True, blank=True)
-    ip = models.GenericIPAddressField(null=True, blank=True)
-    url_redirect = models.ForeignKey(UrlRedirect, models.DO_NOTHING, related_name='logs')
-
-    def __str__(self):
-        return 'log'
